connect -> split
