+++ verticapy/sdk/vertica/dblink.py
@@ -17,7 +17,7 @@ permissions and limitations under the License.
 from verticapy.errors import ConnectionError
 import re
 from verticapy.sql._utils._format import clean_query
-from verticapy.connect.connect import EXTERNAL_CONNECTION
+from verticapy.connect import EXTERNAL_CONNECTION
 
 
 def get_dblink_fun(query: str, symbol: str = "$"):
